backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.3.0
gunicorn==22.0.0